requirements.txt
# Django et extensions
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
drf-spectacular>=0.27.0
python-dotenv>=1.0.0

# Channels pour WebSockets
channels>=4.0.0
daphne>=4.0.0

# PyTorch et dépendances pour le modèle d'IA
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
Pillow>=10.0.0

# Utilitaires
uuid>=1.30
numpy>=1.24.0
